@@ -198,13 +198,17 @@ nodeunit = (function () {
 					: null
 			}
 
-			String.prototype.toJSON = Number.prototype.toJSON = Boolean.prototype.toJSON = function (key) {
-				return this.valueOf()
-			}
+			String.prototype.toJSON =
+				Number.prototype.toJSON =
+				Boolean.prototype.toJSON =
+					function (key) {
+						return this.valueOf()
+					}
 		}
 
 		var cx = /[\u0000\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/g,
-			escapable = /[\\\"\x00-\x1f\x7f-\x9f\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/g,
+			escapable =
+				/[\\\"\x00-\x1f\x7f-\x9f\u00ad\u0600-\u0604\u070f\u17b4\u17b5\u200c-\u200f\u2028-\u202f\u2060-\u206f\ufeff\ufff0-\uffff]/g,
 			gap,
 			indent,
 			meta = {
@@ -313,8 +317,8 @@ nodeunit = (function () {
 							partial.length === 0
 								? '[]'
 								: gap
-								? '[\n' + gap + partial.join(',\n' + gap) + '\n' + mind + ']'
-								: '[' + partial.join(',') + ']'
+									? '[\n' + gap + partial.join(',\n' + gap) + '\n' + mind + ']'
+									: '[' + partial.join(',') + ']'
 						gap = mind
 						return v
 					}
@@ -352,8 +356,8 @@ nodeunit = (function () {
 						partial.length === 0
 							? '{}'
 							: gap
-							? '{\n' + gap + partial.join(',\n' + gap) + '\n' + mind + '}'
-							: '{' + partial.join(',') + '}'
+								? '{\n' + gap + partial.join(',\n' + gap) + '\n' + mind + '}'
+								: '{' + partial.join(',') + '}'
 					gap = mind
 					return v
 			}
@@ -1176,7 +1180,7 @@ nodeunit = (function () {
 				}
 			}
 		}
-	}.call(this))
+	}).call(this)
 	;(function (exports) {
 		/**
 		 * This file is based on the node.js assert module, but with some small
